fileforge/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
Pillow==10.4.0
pypdf==5.0.1
pandas==2.2.3
openpyxl==3.1.5
aiofiles==24.1.0
